requirements.txt
fastapi==0.110.0
uvicorn==0.27.0
python-dotenv==1.0.0
requests==2.31.0
supabase==2.4.3
httpx==0.24.1
gotrue==2.4.2
httpcore==0.17.3